requirements.txt
astroid==2.1.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-stripe==1.2.2
Django==2.0.7
django-allauth==0.36.0
django-braces==1.13.0
django-classy-tags==0.8.0
django-cms==3.6.0
django-crispy-forms==1.7.2
django-excel==0.0.10
django-formtools==2.1
django-heroku==0.3.1
django-import-export==1.1.0
django-model-utils==3.1.2
django-sekizai==0.10.0
django-sslify==0.2.7
django-star-ratings==0.7.0
django-treebeard==4.3
django-use-email-as-username==1.0.1
et-xmlfile==1.0.1
facebook-sdk==3.1.0
googlemaps==3.0.2
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
isort==4.3.4
jdcal==1.4
jsonfield==2.0.2
lazy-object-proxy==1.3.1
lml==0.0.7
mccabe==0.6.1
model-mommy==1.3.2
mysqlclient==1.4.1
oauthlib==2.1.0
odfpy==1.4.0
openpyxl==2.5.12
pendulum==2.0.4
Pillow==5.2.0
pipenv==2018.11.26
protobuf==3.6.1
psycopg2==2.7.7
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
PyJWT==1.7.1
pylint==2.2.2
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
pytzdata==2018.9
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
stripe==2.4.0
swapper==1.1.0
tablib==0.12.1
texttable==1.5.0
unicodecsv==0.14.1
urllib3==1.23
virtualenv==16.4.3
virtualenv-clone==0.5.1
whitenoise==4.1.2
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
